Split shared chroma utilities and orchestrators out of chroma/qam.py
Moves the format-agnostic pieces (chroma_to_u16, the pre-TBC
colour-under filter, and the process_chroma/decode_chroma per-field
orchestrators) verbatim into vhsdecode/chroma/common.py, leaving qam.py
with the burst measurement and heterodyne up-conversion pipeline. The
import graph is a simple DAG: common -> {qam, secam}. get_burst_area
stays in qam.py since the burst machinery there uses it. No behavior
change.
